requirements.txt
pyyaml>=6.0
aiohttp>=3.8.0
